[feature] 강아지 정보 구정

+++ src/main/java/com/example/petshield/domain/Dog.java
@@ -53,5 +53,5 @@ public class Dog extends BaseEntity {
     private Extra extra;
 
     @Column(nullable = false)
-    private Float activityLevel;
+    private Double activityLevel;
 }
